It turns out we don't want all this dll-decl stuff for mingw. So make the #ifdef correct for mscv-but-not-mingw.
I think the right fix would be to take all this logic out and
insert it via a MOE directive, just for the windows versions
of these files.  This is what ctemplate does.  But it's a
more-major change, and I'm looking to get this release out
today, so I'm submitting this (TBR), since it's a much more
minor change.


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2837


git-svn-id: http://gflags.googlecode.com/svn@55 6586e3c6-dcc4-952a-343f-ff74eb82781d